+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [x] Enable async completion
 - [x] Remove *Code Navigation* group from context menu
 - [x] Only show completion for `root` at top of document
-- [ ] Filter completion items already set in same section
+- [x] Support *CompleteWord* command
 - [ ] Add item template for .editorconfig files
 
 Features that have a checkmark are complete and available for